pyproject.toml
[project]
name = "creation-hospital-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "blinker==1.9.0",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "chromadb>=1.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "distro==1.9.0",
    "faiss-cpu==1.13.2",
    "filelock==3.20.3",
    "flask==3.1.2",
    "fsspec==2026.1.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.2",
    "google-api-python-client==2.188.0",
    "google-auth==2.47.0",
    "google-auth-httplib2==0.3.0",
    "google-cloud-aiplatform>=1.133.0",
    "google-cloud-speech==2.36.0",
    "google-cloud-texttospeech==2.34.0",
    "google-genai==1.59.0",
    "google-generativeai==0.8.6",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "grpcio-status==1.71.2",
    "gtts==2.5.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.1",
    "httpx==0.28.1",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "llvmlite==0.46.0",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "mpmath==1.3.0",
    "networkx==3.6.1",
    "numba==0.63.1",
    "numpy==2.3.5",
    "openai-whisper==20250625",
    "packaging==25.0",
    "proto-plus==1.27.0",
    "protobuf==5.29.5",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyparsing==3.3.1",
    "pypdf==6.6.0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "regex==2026.1.15",
    "requests==2.32.5",
    "rich==14.2.0",
    "rsa==4.9.1",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.0",
    "sentence-transformers==5.2.0",
    "setuptools==80.9.0",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "streamlit>=1.53.0",
    "streamlit-mic-recorder>=0.0.8",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "torch==2.9.1",
    "tqdm==4.67.1",
    "transformers==4.57.6",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "websockets==15.0.1",
    "werkzeug==3.1.5",
]
